flood_alarm_api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
numpy
tensorflow
requests
pyngrok